Create encryption_caesarcipher.c
A program used to encrypt text input using a simple Caesar Cipher (shift). The Caesar Cipher uses a substitution method where letters in the alphabet are shifted by a fixed number of spaces (input from the user in the form of a whole number [the key]) to generate a new encoded alphabet. The program takes the key from the user (ex: 4) and a text input (ex: book) and encrypts the text input by shifting every letter in the input by 4 letters. So the example "book" becomes "FSSO".